% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_struct_seir_ode.R
\name{age_struct_seir_ode}
\alias{age_struct_seir_ode}
\title{Age-structured SEIR ODE model of vaccination with 2 doses and delay to protection}
\usage{
age_struct_seir_ode(times, init, params)
}
\arguments{
\item{times}{vector of times}

\item{init}{list of initial states}

\item{params}{list of parameter values}
}
\value{
List of summary results
}
\description{
Age-structured SEIR ODE model of vaccination with 2 doses and delay to protection
}
\keyword{vacamole}
