% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_struct_seir_ode2.R
\name{age_struct_seir_ode2}
\alias{age_struct_seir_ode2}
\title{Age-structured SEIR ODE model of vaccination with 2 dose primary
series with 3 booster doses (3rd, 4th, and 5th doses)}
\usage{
age_struct_seir_ode2(times, init, params)
}
\arguments{
\item{times}{vector of times}

\item{init}{list of initial states}

\item{params}{list of parameter values}
}
\value{
List of summary results
}
\description{
Age-structured SEIR ODE model of vaccination with 2 dose primary
series with 3 booster doses (3rd, 4th, and 5th doses)
}
\keyword{vacamole}
