% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ve_w_waning.R
\name{calc_ve_w_waning}
\alias{calc_ve_w_waning}
\title{Calculate amount of vaccine waning}
\usage{
calc_ve_w_waning(vac_rate, ve_val, waning)
}
\arguments{
\item{vac_rate}{rate of vaccination at each timepoint}

\item{ve_val}{value of VE}

\item{waning}{vector of amount of waning over time}
}
\value{
matrix of VE values with waning in each age group (columns) at each
time point (rows)
}
\description{
Calculate amount of vaccine waning
}
\keyword{vacamole}
