% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_waning.R
\name{calc_waning}
\alias{calc_waning}
\title{Calculate vaccine efficacy waning}
\usage{
calc_waning(prop, time_points, k, t0)
}
\description{
Based on the weighted overage of the VE, daily vaccination rate,
 and time since vaccination
}
