% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_contact_matrix.R
\name{choose_contact_matrix}
\alias{choose_contact_matrix}
\title{Determine contact matrix based on thresholds of cases or IC admissions}
\usage{
choose_contact_matrix(
  times,
  params,
  criteria,
  flag_relaxed,
  flag_very_relaxed,
  flag_normal,
  keep_fixed
)
}
\arguments{
\item{params}{list of parameter values}

\item{criteria}{criteria by which to change contact matrix. User can choose between a number of
cases per day or a number of IC admissions.}

\item{flag_relaxed}{an integer value that becomes non-zero (positive) when the contact matrix consistent with
"relaxed" non-pharmaceutical interventions is being used.}

\item{flag_very_relaxed}{an integer value that becomes non-zero (positive) when the contact matrix consistent with
"very relaxed" non-pharmaceutical interventions is being used.}

\item{flag_normal}{an integer value that becomes non-zero (positive) when the contact matrix consistent with
normal (pre-pandemic) contact patterns is being used.}

\item{keep_fixed}{logical. if TRUE the contact matrix stays fixed over the
entire simulation period}
}
\value{
List of summary results
}
\description{
Determine contact matrix based on thresholds of cases or IC admissions
}
\keyword{vacamole}
