% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_contact_matrices.R
\name{convert_contact_matrices}
\alias{convert_contact_matrices}
\title{Convert contact matrices to transmission matrices}
\usage{
convert_contact_matrices(x)
}
\arguments{
\item{x}{data frame with contacts}
}
\value{
list of contact matrices
}
\description{
Convert contact matrices to transmission matrices
}
\keyword{vacamole}
