% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_data_func.R
\name{fit_to_data_func}
\alias{fit_to_data_func}
\title{Fit to data}
\usage{
fit_to_data_func(
  breakpoints,
  params,
  init,
  fit_pars,
  case_data,
  contact_matrices,
  vac_info,
  save_output_to_file = TRUE,
  path_out = NULL,
  ...
)
}
\value{
list with output XX add more details XX
}
\description{
Fit to data
}
\keyword{vacamole}
