% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sim_func_wrap.R
\name{forward_sim_func_wrap}
\alias{forward_sim_func_wrap}
\title{Wrapper function to perform forward simulations}
\usage{
forward_sim_func_wrap(
  params,
  start_date,
  end_date,
  init_cond,
  beta_m,
  vac_inputs,
  beta_c,
  beta_d,
  t_normal,
  contact_matrices,
  tag
)
}
\arguments{
\item{params}{List of model input parameters.}

\item{start_date}{Start date of simulations (character string in YYYY-MM-DD format)}

\item{end_date}{End date of simulations (character string in YYYY-MM-DD format)}

\item{init_cond}{Named vector of initial conditions for model compartments}

\item{beta_m}{Transmission rate mean.}

\item{vac_inputs}{Vaccination inputs for model. This should be the output of convert_vac_schedule().}

\item{beta_c}{Transmission rate when contact matrix changes to pre-pandemic (time > t_normal)}

\item{beta_d}{Transmission rate draws for determining uncertainty around the mean transmission rate}

\item{t_normal}{Time point to force model to switch to pre-pandemic contact matrix}

\item{contact_matrices}{List of contact matrices}

\item{tag}{Character string of name of output file}
}
\value{
Data frame of model outputs for each simulation
}
\description{
Wrapper function to perform forward simulations
}
\keyword{vacamole}
