% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foi.R
\name{get_foi}
\alias{get_foi}
\title{Calculate force of infection}
\usage{
get_foi(dat, params, vac_inputs)
}
\arguments{
\item{dat}{data frame of states}

\item{params}{list of parameter values}

\item{contact_matrix}{contact matrix}
}
\value{
matrix of force of infection in each age group (columns) at each
time point (rows)
}
\description{
Calculate force of infection
}
\keyword{vacamole}
