% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transmission_matrix.R
\name{get_transmission_matrix}
\alias{get_transmission_matrix}
\title{Convert contact matrix to transmission matrix}
\usage{
get_transmission_matrix(x, contact_mat)
}
\arguments{
\item{x}{vector of relative susceptibility and infectiousnes for each age group}

\item{contact_mat}{contact matrix}
}
\value{
transmission matrix
}
\description{
Convert contact matrix to transmission matrix
}
\keyword{vacamole}
