% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_func.R
\name{likelihood_func}
\alias{likelihood_func}
\title{Likelihood function for fitting SEIR model to daily case data}
\usage{
likelihood_func(x, t, data, params, init, stochastic = FALSE)
}
\arguments{
\item{x}{vector of parameters to fit}

\item{t}{vector of time points}

\item{data}{real data to fit model to}

\item{params}{list of parameters values for input into SEIR model}

\item{init}{named vector of initial conditions for each model compartment}

\item{stochastic}{logical, if TRUE, a stochastic model is fit to the data
otherwise a deterministic model is used.}
}
\description{
Likelihood function for fitting SEIR model to daily case data
}
\keyword{vacamole}
