% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run_wrapper.R
\name{model_run_wrapper}
\alias{model_run_wrapper}
\title{Function wrapper to run model for multiple time windows with set of 
transmission rates}
\usage{
model_run_wrapper(
  breakpoints,
  beta_values,
  init_conditions,
  params,
  mle = TRUE
)
}
\arguments{
\item{breakpoints}{list of breakpoints}

\item{beta_values}{vector of beta values}

\item{init_conditions}{vector of initial conditions}

\item{params}{vector of model input parameters}

\item{mle}{logical, if TRUE the current run is the maximum likelihood estimate}
}
\description{
Function wrapper to run model for multiple time windows with set of 
transmission rates
}
\keyword{vacamole}
