% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rmultinom.R
\name{my_rmultinom}
\alias{my_rmultinom}
\title{Postprocess output from age-structured SEIR ODE model of vaccination 
with 2 doses and delay to protection}
\usage{
my_rmultinom(x)
}
\arguments{
\item{x}{vector of probabilities}
}
\value{
List of summary results
}
\description{
Postprocess output from age-structured SEIR ODE model of vaccination 
with 2 doses and delay to protection
}
\keyword{vacamole}
