% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_age_struct_model_output2.R
\name{postprocess_age_struct_model_output2}
\alias{postprocess_age_struct_model_output2}
\title{Postprocess output from age-structured SEIR ODE model of vaccination
with 2 doses and delay to protection}
\usage{
postprocess_age_struct_model_output2(dat)
}
\arguments{
\item{dat}{output from seir model as a data frame}
}
\value{
List of summary results
}
\description{
Postprocess output from age-structured SEIR ODE model of vaccination
with 2 doses and delay to protection
}
\keyword{vacamole}
