% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Dynamic agent-based model of infection and vaccination}
\usage{
run_model(
  pop_size,
  age_dist,
  hh_info,
  n_postcodes,
  prop_inf_by_age,
  prop_rec_by_age,
  t_max,
  vac_mech,
  contact_mat,
  prop_in_pc,
  sigma,
  beta,
  tau,
  ve,
  prop_asympt_vec,
  prop_severe,
  prop_hosp,
  prop_non_hosp_death,
  prop_ICU,
  prop_hosp_death,
  prop_ICU_death,
  vaccines_per_day,
  perc_vac_hesitant,
  perc_high_risk,
  time_in_hh_isolation,
  time_btw_doses,
  time_to_protection,
  vac_age_group_order
)
}
\arguments{
\item{pop_size}{Number of individuals in the population.}

\item{age_dist}{Age distribution: proportion of individuals in each age group}

\item{n_postcodes}{number of distict geographic regions to divide population into
if 0 then there will be no household structure}

\item{prop_inf_by_age}{vector of proportion of each age group infected at the start of the simulation}

\item{t_max}{integer of final time step (days)}

\item{prop_in_pc}{proportion of contacts that are within an individual's postcode}

\item{sigma}{matrix of susceptibilities for each age group
@param beta matrix of age-stratified transmission rates}

\item{tau}{relative transmissibility of asymptomatic infection}

\item{ve}{vaccine efficacy}

\item{hh_info_mat}{data frame with household distribution information}

\item{vaccine_mech}{integer denoting vaccine mechanism:
1 - vaccine reduces susceptibility
2 - vaccine reduces the probability of becoming symptomatic
3 - vaccine protects against severe symptoms}

\item{contact_matrix}{matrix of daily contacts for each age group (should be of dimension length(age_dist) x length(age_dist))}

\item{prop_asymp_vec}{vector of proportions of asymptomatic infections by age group}
}
\value{
vectors of XX
}
\description{
This function initializes the population before running the model.
}
\keyword{vac-a-mole}
