% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_ode.R
\name{seir_ode}
\alias{seir_ode}
\title{SEIR ODE model of vaccination with 2 doses and delay to protection}
\usage{
seir_ode(times, init, params)
}
\arguments{
\item{times}{vector of times}

\item{init}{list of inititial states}

\item{params}{list of parameter values}
}
\value{
List of summary results
}
\description{
SEIR ODE model of vaccination with 2 doses and delay to protection
}
\keyword{vacamole}
