% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_age_struct_seir_ode.R
\name{stochastic_age_struct_seir_ode}
\alias{stochastic_age_struct_seir_ode}
\title{Stochastic Age-structured SEIR ODE model of vaccination with 2 doses and delay to protection}
\usage{
stochastic_age_struct_seir_ode(times, init, params)
}
\arguments{
\item{times}{vector of times}

\item{init}{list of initial states}

\item{params}{list of parameter values}
}
\value{
List of summary results
}
\description{
Stochastic Age-structured SEIR ODE model of vaccination with 2 doses and delay to protection
}
\keyword{vacamole}
