% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise output from SEIR model for plotting}
\usage{
summarise_results(seir_output, params, start_date, times, vac_inputs)
}
\arguments{
\item{seir_output}{list of dataframes from postprocessing}

\item{params}{list of parameter values}

\item{start_date}{calendar date of start of simulation}

\item{times}{vector of time points}
}
\value{
List of summary results
}
\description{
Summarise output from SEIR model for plotting
}
\keyword{vacamole}
