% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Convert data from wide to long}
\usage{
wide_to_long(x, times)
}
\arguments{
\item{x}{data frame to convert}

\item{times}{vector of time points. should be the same length as
nrow(x)}
}
\value{
data frame in long format
}
\description{
Convert data from wide to long
}
\keyword{vacamole}
