% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wrangle_results}
\alias{wrangle_results}
\title{Post-process raw results from model run}
\usage{
wrangle_results(x)
}
\arguments{
\item{x}{output from model run}
}
\value{
data frame in long format
}
\description{
Post-process raw results from model run
}
\keyword{vacamole}
