% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vac_schedule.R
\name{convert_vac_schedule}
\alias{convert_vac_schedule}
\title{Convert cumulative vaccination schedule to non-cumulative}
\usage{
convert_vac_schedule(
  vac_schedule,
  ve_pars,
  wane = FALSE,
  k_inf = 0.012,
  k_sev = 0.006,
  t0 = 365
)
}
\arguments{
\item{vac_schedule}{a data frame that the proportion of the population who
receives vaccines at each time point. Rows are time points, columns are
<vaccine type>_<dose>_<age group>. For example, the first column is the
proportion of individuals in age group 1 who receive dose 1 of the first
vaccine type.
The function assumes 9 or 10 age groups (1, .., 10) and four vaccine types
(pf, mo, az, ja), each with a 2-dose regimen (d1, d2).}

\item{ve_pars}{a data frame with VE estimates by vaccine product, dose,
outcome, and age group.
The data frame should have the following columns: vac_product, dose,
age_group, outcome, variant, delay, ve.}

\item{wane}{logical, if TRUE vaccine effectiveness wanes by a logistic
function parameterized by arguments k and t0.}

\item{k_inf}{logistic growth rate for waning curve against infection}

\item{k_sev}{logistic growth rate for waning curve against severe disease}

\item{t0}{the time point at the midpoint of the logistic curve (where 50\%
waning occurs)}
}
\value{
data frame of vaccination rate by day, dose, vaccine product, and age
group and weighted VE and delay to protection by day, dose, age group, and
outcome
}
\description{
Convert cumulative vaccination schedule to non-cumulative
}
\keyword{vacamole}
