% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foi.R
\name{get_foi}
\alias{get_foi}
\title{Calculate force of infection}
\usage{
get_foi(x, y1, y2, y3, y4, y5, beta, contact_mat, times)
}
\arguments{
\item{x}{list of data frames for each state from SEIR model}

\item{y1}{data frame of protection against transmission due to vaccination from dose 1 at each time point}

\item{y2}{data frame of protection against transmission due to vaccination from dose 2 at each time point}

\item{y3}{data frame of protection against transmission due to vaccination from dose 3 at each time point}

\item{y4}{data frame of protection against transmission due to vaccination from dose 4 at each time point}

\item{y5}{data frame of protection against transmission due to vaccination from dose 5 at each time point}

\item{beta}{vector of values of transmission rate for each time point}

\item{contact_mat}{contact matrix}

\item{times}{vector of time points that correspond to the rows of x}
}
\value{
matrix of force of infection in each age group (columns) at each
time point (rows)
}
\description{
Calculate force of infection
}
\keyword{vacamole}
